% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca_explained_variance}
\alias{pca_explained_variance}
\title{Explained variance of each principal component}
\usage{
pca_explained_variance(pca)
}
\arguments{
\item{pca}{PCA object returned by pca_fit}
}
\value{
a vector of each principal component's explained variance
}
\description{
Explained variance of each principal component
}
