% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca_transform}
\alias{pca_transform}
\title{Returns the principal components of x}
\usage{
pca_transform(x, pca, ncomp = NA)
}
\arguments{
\item{x}{input matrix}

\item{pca}{PCA object returned by pca_fit}

\item{ncomp}{number of PCs to return or NA for all of them}
}
\value{
x transformed to its principal components
}
\description{
Returns the principal components of x
}
