% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{scale_fit}
\alias{scale_fit}
\title{Scales the columns of the dataset}
\usage{
scale_fit(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{dataset with observations in rows, variables in columns}

\item{center}{whether to center the columns to mean 0}

\item{scale}{whether to scale columns to variance 1}
}
\value{
parameters for scaling the given dataset which can be applied to other data
}
\description{
Scales the columns of the dataset
}
