% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{scale_transform}
\alias{scale_transform}
\title{Transforms a dataset using the given scaling parameters}
\usage{
scale_transform(x, scal)
}
\arguments{
\item{x}{dataset with observations in rows, variables in columns}

\item{scal}{the scaling object}
}
\value{
the normalized data
}
\description{
Transforms a dataset using the given scaling parameters
}
