% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{subset_cols}
\alias{subset_cols}
\title{Subsets the first n columns of x}
\usage{
subset_cols(x, n)
}
\arguments{
\item{x}{matrix or data frame}

\item{n}{number of columns to subset}
}
\value{
a matrix with the first n columns of x
}
\description{
This may be needed when n = 1 so that R doesn't coerce
the result as a numeric vector
}
