% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{scale_untransform}
\alias{scale_untransform}
\title{Returns a normalized dataset to its original scale}
\usage{
scale_untransform(scal, x)
}
\arguments{
\item{scal}{the scaling object}

\item{x}{the data normalized by columns}
}
\value{
the data in its original scale
}
\description{
Returns a normalized dataset to its original scale
}
